[Concurrency] restructuring
