+++ src/main/java/com/epam/jmp/concurrency/files/FilesReader.java
@@ -1,4 +1,4 @@
-package com.epam.jmp.concurrency.impl;
+package com.epam.jmp.concurrency.files;
 
 import java.nio.file.Path;
 import java.util.List;
