--- src/main/java/com/epam/jmp/concurrency/impl/NewsService.java
@@ -1,7 +1,0 @@
-package com.epam.jmp.concurrency.impl;
-
-public interface NewsService {
-
-	boolean upsertNews(News news);
-
-}
