+++ src/main/java/com/epam/jmp/concurrency/job/ReaderJob.java
@@ -1,4 +1,4 @@
-package com.epam.jmp.concurrency.impl;
+package com.epam.jmp.concurrency.job;
 
 import static com.epam.jmp.util.Config.READ_TIMEOUT;
 
@@ -9,6 +9,7 @@ import java.util.concurrent.TimeUnit;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
+import com.epam.jmp.concurrency.files.FilesReader;
 import com.epam.jmp.util.Config;
 
 public class ReaderJob implements Runnable {
