+++ src/main/java/com/epam/jmp/concurrency/service/LocalNewsService.java
@@ -1,7 +1,9 @@
-package com.epam.jmp.concurrency.impl;
+package com.epam.jmp.concurrency.service;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
+
+import com.epam.jmp.concurrency.model.News;
 
 public class LocalNewsService implements NewsService {
 	private Logger LOG = LogManager.getLogger(LocalNewsService.class);
