+++ src/main/java/com/epam/jmp/concurrency/service/NewsService.java
@@ -1,0 +1,9 @@
+package com.epam.jmp.concurrency.service;
+
+import com.epam.jmp.concurrency.model.News;
+
+public interface NewsService {
+
+	boolean upsertNews(News news);
+
+}
